# pkg/proto/xmtpv4/message_api/misbehavior_api.pb.gw.go
// Code generated by protoc-gen-grpc-gateway. DO NOT EDIT.
// source: xmtpv4/message_api/misbehavior_api.proto

/*
Package message_api is a reverse proxy.

It translates gRPC into RESTful JSON APIs.
*/
package message_api

import (
	"context"
	"io"
	"net/http"

	"github.com/grpc-ecosystem/grpc-gateway/v2/runtime"
	"github.com/grpc-ecosystem/grpc-gateway/v2/utilities"
	"google.golang.org/grpc"
	"google.golang.org/grpc/codes"
	"google.golang.org/grpc/grpclog"
	"google.golang.org/grpc/metadata"
	"google.golang.org/grpc/status"
	"google.golang.org/protobuf/proto"
)

// Suppress "imported and not used" errors
var _ codes.Code
var _ io.Reader
var _ status.Status
var _ = runtime.String
var _ = utilities.NewDoubleArray
var _ = metadata.Join

func request_MisbehaviorApi_SubmitMisbehaviorReport_0(ctx context.Context, marshaler runtime.Marshaler, client MisbehaviorApiClient, req *http.Request, pathParams map[string]string) (proto.Message, runtime.ServerMetadata, error) {
	var protoReq SubmitMisbehaviorReportRequest
	var metadata runtime.ServerMetadata

	if err := marshaler.NewDecoder(req.Body).Decode(&protoReq); err != nil && err != io.EOF {
		return nil, metadata, status.Errorf(codes.InvalidArgument, "%v", err)
	}

	msg, err := client.SubmitMisbehaviorReport(ctx, &protoReq, grpc.Header(&metadata.HeaderMD), grpc.Trailer(&metadata.TrailerMD))
	return msg, metadata, err

}

func local_request_MisbehaviorApi_SubmitMisbehaviorReport_0(ctx context.Context, marshaler runtime.Marshaler, server MisbehaviorApiServer, req *http.Request, pathParams map[string]string) (proto.Message, runtime.ServerMetadata, error) {
	var protoReq SubmitMisbehaviorReportRequest
	var metadata runtime.ServerMetadata

	if err := marshaler.NewDecoder(req.Body).Decode(&protoReq); err != nil && err != io.EOF {
		return nil, metadata, status.Errorf(codes.InvalidArgument, "%v", err)
	}

	msg, err := server.SubmitMisbehaviorReport(ctx, &protoReq)
	return msg, metadata, err

}

func request_MisbehaviorApi_QueryMisbehaviorReports_0(ctx context.Context, marshaler runtime.Marshaler, client MisbehaviorApiClient, req *http.Request, pathParams map[string]string) (proto.Message, runtime.ServerMetadata, error) {
	var protoReq QueryMisbehaviorReportsRequest
	var metadata runtime.ServerMetadata

	if err := marshaler.NewDecoder(req.Body).Decode(&protoReq); err != nil && err != io.EOF {
		return nil, metadata, status.Errorf(codes.InvalidArgument, "%v", err)
	}

	msg, err := client.QueryMisbehaviorReports(ctx, &protoReq, grpc.Header(&metadata.HeaderMD), grpc.Trailer(&metadata.TrailerMD))
	return msg, metadata, err

}

func local_request_MisbehaviorApi_QueryMisbehaviorReports_0(ctx context.Context, marshaler runtime.Marshaler, server MisbehaviorApiServer, req *http.Request, pathParams map[string]string) (proto.Message, runtime.ServerMetadata, error) {
	var protoReq QueryMisbehaviorReportsRequest
	var metadata runtime.ServerMetadata

	if err := marshaler.NewDecoder(req.Body).Decode(&protoReq); err != nil && err != io.EOF {
		return nil, metadata, status.Errorf(codes.InvalidArgument, "%v", err)
	}

	msg, err := server.QueryMisbehaviorReports(ctx, &protoReq)
	return msg, metadata, err

}

// RegisterMisbehaviorApiHandlerServer registers the http handlers for service MisbehaviorApi to "mux".
// UnaryRPC     :call MisbehaviorApiServer directly.
// StreamingRPC :currently unsupported pending https://github.com/grpc/grpc-go/issues/906.
// Note that using this registration option will cause many gRPC library features to stop working. Consider using RegisterMisbehaviorApiHandlerFromEndpoint instead.
func RegisterMisbehaviorApiHandlerServer(ctx context.Context, mux *runtime.ServeMux, server MisbehaviorApiServer) error {

	mux.Handle("POST", pattern_MisbehaviorApi_SubmitMisbehaviorReport_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
		ctx, cancel := context.WithCancel(req.Context())
		defer cancel()
		var stream runtime.ServerTransportStream
		ctx = grpc.NewContextWithServerTransportStream(ctx, &stream)
		inboundMarshaler, outboundMarshaler := runtime.MarshalerForRequest(mux, req)
		var err error
		var annotatedContext context.Context
		annotatedContext, err = runtime.AnnotateIncomingContext(ctx, mux, req, "/xmtp.xmtpv4.message_api.MisbehaviorApi/SubmitMisbehaviorReport", runtime.WithHTTPPathPattern("/mls/v2/submit-misbehavior-report"))
		if err != nil {
			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
			return
		}
		resp, md, err := local_request_MisbehaviorApi_SubmitMisbehaviorReport_0(annotatedContext, inboundMarshaler, server, req, pathParams)
		md.HeaderMD, md.TrailerMD = metadata.Join(md.HeaderMD, stream.Header()), metadata.Join(md.TrailerMD, stream.Trailer())
		annotatedContext = runtime.NewServerMetadataContext(annotatedContext, md)
		if err != nil {
			runtime.HTTPError(annotatedContext, mux, outboundMarshaler, w, req, err)
			return
		}

		forward_MisbehaviorApi_SubmitMisbehaviorReport_0(annotatedContext, mux, outboundMarshaler, w, req, resp, mux.GetForwardResponseOptions()...)

	})

	mux.Handle("POST", pattern_MisbehaviorApi_QueryMisbehaviorReports_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
		ctx, cancel := context.WithCancel(req.Context())
		defer cancel()
		var stream runtime.ServerTransportStream
		ctx = grpc.NewContextWithServerTransportStream(ctx, &stream)
		inboundMarshaler, outboundMarshaler := runtime.MarshalerForRequest(mux, req)
		var err error
		var annotatedContext context.Context
		annotatedContext, err = runtime.AnnotateIncomingContext(ctx, mux, req, "/xmtp.xmtpv4.message_api.MisbehaviorApi/QueryMisbehaviorReports", runtime.WithHTTPPathPattern("/mls/v2/query-misbehavior-reports"))
		if err != nil {
			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
			return
		}
		resp, md, err := local_request_MisbehaviorApi_QueryMisbehaviorReports_0(annotatedContext, inboundMarshaler, server, req, pathParams)
		md.HeaderMD, md.TrailerMD = metadata.Join(md.HeaderMD, stream.Header()), metadata.Join(md.TrailerMD, stream.Trailer())
		annotatedContext = runtime.NewServerMetadataContext(annotatedContext, md)
		if err != nil {
			runtime.HTTPError(annotatedContext, mux, outboundMarshaler, w, req, err)
			return
		}

		forward_MisbehaviorApi_QueryMisbehaviorReports_0(annotatedContext, mux, outboundMarshaler, w, req, resp, mux.GetForwardResponseOptions()...)

	})

	return nil
}

// RegisterMisbehaviorApiHandlerFromEndpoint is same as RegisterMisbehaviorApiHandler but
// automatically dials to "endpoint" and closes the connection when "ctx" gets done.
func RegisterMisbehaviorApiHandlerFromEndpoint(ctx context.Context, mux *runtime.ServeMux, endpoint string, opts []grpc.DialOption) (err error) {
	conn, err := grpc.DialContext(ctx, endpoint, opts...)
	if err != nil {
		return err
	}
	defer func() {
		if err != nil {
			if cerr := conn.Close(); cerr != nil {
				grpclog.Infof("Failed to close conn to %s: %v", endpoint, cerr)
			}
			return
		}
		go func() {
			<-ctx.Done()
			if cerr := conn.Close(); cerr != nil {
				grpclog.Infof("Failed to close conn to %s: %v", endpoint, cerr)
			}
		}()
	}()

	return RegisterMisbehaviorApiHandler(ctx, mux, conn)
}

// RegisterMisbehaviorApiHandler registers the http handlers for service MisbehaviorApi to "mux".
// The handlers forward requests to the grpc endpoint over "conn".
func RegisterMisbehaviorApiHandler(ctx context.Context, mux *runtime.ServeMux, conn *grpc.ClientConn) error {
	return RegisterMisbehaviorApiHandlerClient(ctx, mux, NewMisbehaviorApiClient(conn))
}

// RegisterMisbehaviorApiHandlerClient registers the http handlers for service MisbehaviorApi
// to "mux". The handlers forward requests to the grpc endpoint over the given implementation of "MisbehaviorApiClient".
// Note: the gRPC framework executes interceptors within the gRPC handler. If the passed in "MisbehaviorApiClient"
// doesn't go through the normal gRPC flow (creating a gRPC client etc.) then it will be up to the passed in
// "MisbehaviorApiClient" to call the correct interceptors.
func RegisterMisbehaviorApiHandlerClient(ctx context.Context, mux *runtime.ServeMux, client MisbehaviorApiClient) error {

	mux.Handle("POST", pattern_MisbehaviorApi_SubmitMisbehaviorReport_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
		ctx, cancel := context.WithCancel(req.Context())
		defer cancel()
		inboundMarshaler, outboundMarshaler := runtime.MarshalerForRequest(mux, req)
		var err error
		var annotatedContext context.Context
		annotatedContext, err = runtime.AnnotateContext(ctx, mux, req, "/xmtp.xmtpv4.message_api.MisbehaviorApi/SubmitMisbehaviorReport", runtime.WithHTTPPathPattern("/mls/v2/submit-misbehavior-report"))
		if err != nil {
			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
			return
		}
		resp, md, err := request_MisbehaviorApi_SubmitMisbehaviorReport_0(annotatedContext, inboundMarshaler, client, req, pathParams)
		annotatedContext = runtime.NewServerMetadataContext(annotatedContext, md)
		if err != nil {
			runtime.HTTPError(annotatedContext, mux, outboundMarshaler, w, req, err)
			return
		}

		forward_MisbehaviorApi_SubmitMisbehaviorReport_0(annotatedContext, mux, outboundMarshaler, w, req, resp, mux.GetForwardResponseOptions()...)

	})

	mux.Handle("POST", pattern_MisbehaviorApi_QueryMisbehaviorReports_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
		ctx, cancel := context.WithCancel(req.Context())
		defer cancel()
		inboundMarshaler, outboundMarshaler := runtime.MarshalerForRequest(mux, req)
		var err error
		var annotatedContext context.Context
		annotatedContext, err = runtime.AnnotateContext(ctx, mux, req, "/xmtp.xmtpv4.message_api.MisbehaviorApi/QueryMisbehaviorReports", runtime.WithHTTPPathPattern("/mls/v2/query-misbehavior-reports"))
		if err != nil {
			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
			return
		}
		resp, md, err := request_MisbehaviorApi_QueryMisbehaviorReports_0(annotatedContext, inboundMarshaler, client, req, pathParams)
		annotatedContext = runtime.NewServerMetadataContext(annotatedContext, md)
		if err != nil {
			runtime.HTTPError(annotatedContext, mux, outboundMarshaler, w, req, err)
			return
		}

		forward_MisbehaviorApi_QueryMisbehaviorReports_0(annotatedContext, mux, outboundMarshaler, w, req, resp, mux.GetForwardResponseOptions()...)

	})

	return nil
}

var (
	pattern_MisbehaviorApi_SubmitMisbehaviorReport_0 = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0, 2, 1, 2, 2}, []string{"mls", "v2", "submit-misbehavior-report"}, ""))

	pattern_MisbehaviorApi_QueryMisbehaviorReports_0 = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0, 2, 1, 2, 2}, []string{"mls", "v2", "query-misbehavior-reports"}, ""))
)

var (
	forward_MisbehaviorApi_SubmitMisbehaviorReport_0 = runtime.ForwardResponseMessage

	forward_MisbehaviorApi_QueryMisbehaviorReports_0 = runtime.ForwardResponseMessage
)